% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R
\name{df_lp_games}
\alias{df_lp_games}
\alias{df_martian_chess}
\title{Data frames of starting diagrams for various games}
\usage{
df_martian_chess(n_players = 2, seed = NULL)
}
\arguments{
\item{n_players}{Number of players}

\item{seed}{Random seed}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
}
\details{
Here is some more information about the various games:
 \describe{
 \item{Martian Chess}{An abstract by Andrew Looney.
                     See \url{https://www.looneylabs.com/rules/martian-chess}.}
}
}
