% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R
\name{icehouse_games}
\alias{icehouse_games}
\title{Data frames with game info}
\usage{
icehouse_games()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with character columns "game", "methods", "comment", and "url"
for game name, setup function name(s), possible comment, and url for more information.
}
\description{
\code{icehouse_games()} contain information about the games
whose setups are provided by this package.
}
\examples{
head(icehouse_games())
nrow(icehouse_games())
}
