% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{icehouse_parser}
\alias{icehouse_parser}
\alias{looney_pyramid_parser}
\title{Icehouse movetext parser}
\usage{
icehouse_parser(movetext = character(), metadata = list(), ...)

looney_pyramid_parser(movetext = character(), metadata = list(), ...)
}
\arguments{
\item{movetext}{A named character vector of move text.}

\item{metadata}{A named list of metadata}

\item{...}{Passed to \code{\link[ppn:ppn_parsers]{ppn::default_parser()}}.}
}
\value{
A named list of data frames with parsed game states.
}
\description{
\code{icehouse_parser()} provides a parser to be used with PPN movetext.
\code{looney_pyramid_parser()} is an alias.
}
\examples{
if (requireNamespace("ppn", quietly = TRUE)) {
  l <- ppn::read_ppn(system.file("ppn/martian-chess.ppn", package = "piecenikr"),
                     parse = FALSE)[[1L]]
  game <- icehouse_parser(l$movetext, l$metadata)
}
}
