% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icehouse_pieces.R
\name{icehouse_pyramids}
\alias{icehouse_pyramids}
\alias{icehouse_none}
\alias{icehouse_setup_by_name}
\title{Generate icehouse pieces}
\usage{
icehouse_pyramids(
  ...,
  side = "top",
  piece_side = paste0("pyramid_", side),
  suit = rep(1:5, each = 3L),
  rank = rep.int(1:3, 5L),
  x = as.double(rep.int(1:5, 3L)),
  y = as.double(rep(3:1, each = 5L)),
  angle = 0,
  length.out = NA_integer_
)

icehouse_none()

icehouse_setup_by_name(
  name,
  ...,
  getter = function(x) get(x, envir = getNamespace("piecenikr"))
)
}
\arguments{
\item{...}{Should be left empty.}

\item{side}{Normally either "top" (default) or "face".}

\item{piece_side}{If \code{"board_face"} a checkered board (as in American Checkers).
If \code{"board_back"} a lined board (as in Turkish Checkers).}

\item{suit}{Suit value (color) of board and pieces.
\code{1L} is "red", \code{2L} is "black", \code{3L} is "green", \code{4L} is "blue", \code{5L} is "yellow", and \code{6L} is "white".
Will be coerced by \code{\link[ppdf:piece_suit]{piece_suit()}}.}

\item{rank}{Number of pips on the icehouse pyramid.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{angle}{Rotation of bit/board (numeric vector of degrees, counter-clockwise).
Will be coerced by \code{\link[ppdf:piece_angle]{piece_angle()}}.}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}

\item{name}{Game name.  Will be normalized by \code{\link[ppdf:normalize_name]{normalize_name()}}.}

\item{getter}{Function with a single function name argument to use.
By default will just look for functions in this package but
a developer may want to consider \code{\link[=dynGet]{dynGet()}} or a
wrapper around \code{\link[=get]{get()}} with a custom \code{envir} argument.}
}
\description{
\code{icehouse_pyramids()} generates a data frame of icehouse pyramids,
\code{looney_pyramids()} is an alias.
\code{icehouse_none()} generates a zero-row data frame.
\code{icehouse_setup_by_name()} generates a data frame from a given icehouse game name.
}
