% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R
\name{icehouse_setups}
\alias{icehouse_setups}
\alias{icehouse_martian_chess}
\title{Data frames of starting diagrams for various games}
\usage{
icehouse_martian_chess(n_players = 2, seed = NULL)
}
\arguments{
\item{n_players}{Number of players}

\item{seed}{Random seed}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 1 games playable with the Looney Pyramids game system.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various other games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Martian Chess \tab \url{https://www.looneylabs.com/rules/martian-chess} \cr
}
}
