% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icehouse.R
\name{looneyPyramidGrob}
\alias{looneyPyramidGrob}
\title{Basic grob function for Looney Pyramids}
\usage{
looneyPyramidGrob(piece_side, suit, rank, cfg = pp_cfg())
}
\arguments{
\item{piece_side}{Piece side of the piece}

\item{suit}{Suit of the piece}

\item{rank}{Rank of the piece}

\item{cfg}{Piecepackr configuration object}
}
\description{
\code{looneyPyramidGrob} provides a basic grob function for the sides Looney Pyramids.
}
