% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icehouse_pieces.R
\name{icehouse_pyramids}
\alias{icehouse_pyramids}
\alias{icehouse_none}
\title{Generate icehouse pieces}
\usage{
icehouse_pyramids(
  ...,
  side = "top",
  piece_side = paste0("pyramid_", side),
  suit = rep(1:5, each = 3L),
  rank = rep.int(1:3, 5L),
  x = as.double(rep.int(1:5, 3L)),
  y = as.double(rep(3:1, each = 5L)),
  angle = 0,
  length.out = NA_integer_
)

icehouse_none()
}
\arguments{
\item{...}{Should be left empty.}

\item{side}{Normally either "top" (default) or "face".}

\item{piece_side}{If \code{"board_face"} a checkered board (as in American Checkers).
If \code{"board_back"} a lined board (as in Turkish Checkers).}

\item{suit}{Suit value (color) of board.  \code{2L} is "black" and \code{3L} is "green".}

\item{rank}{Number of pips on the icehouse pyramid.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{angle}{Angle of board in degrees.
Italian Checkers rotates its board 90 degrees compared to other checkers variants.}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}
}
\description{
\code{icehouse_pyramids()} generates a data frame of icehouse pyramids.
\code{icehouse_none()} generates a zero-row data frame.
}
