% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icehouse.R
\name{looney_pyramids}
\alias{looney_pyramids}
\title{Configuration objects for drawing Looney Pyramids}
\usage{
looney_pyramids()
}
\value{
Returns a list with an \code{icehouse_pieces} configuration for drawing Looney Pyramids.
}
\description{
\code{looney_pyramids} provides a basic configuration object
for drawing Looney Pyramids.
}
\details{
Looney Pyramids aka Icehouse pieces is a game system invented by Andrew Looney.
See \url{https://www.looneylabs.com/looney-pyramids}.
}
