% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icehouse.R
\docType{data}
\name{looney_pyramids}
\alias{looney_pyramids}
\title{Looney Pyramids configuration object}
\format{An object of class \code{pp_cfg} (inherits from \code{R6}) of length 36.}
\usage{
looney_pyramids
}
\description{
\code{looney_pyramids} provides a basic configuration object
for drawing Looney Pyramids
}
\details{
Looney Pyramids aka Icehouse pieces is a game system invented by Andrew Looney.
See \url{https://www.looneylabs.com/looney-pyramids}.
}
\keyword{datasets}
