% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecenikr-deprecated.R
\name{piecenikr-deprecated}
\alias{piecenikr-deprecated}
\alias{df_martian_chess}
\title{Deprecated functions}
\usage{
df_martian_chess(n_players = 2, seed = NULL)
}
\arguments{
\item{n_players}{Number of players}

\item{seed}{Random seed}
}
\description{
These functions are Deprecated in this release of piecenikr,
they will be marked as Defunct and removed in a future version.
}
